+++ src/main/java/com/menezesmarlon/drogaria/domain/Categoria.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class Categoria {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer codigo;
+    private Short codigo;
 
     @Column(length = 50, nullable = false, unique = true)
     private String nome;
