+++ src/main/java/com/menezesmarlon/drogaria/repository/CategoriaRepository.java
@@ -3,6 +3,6 @@ package com.menezesmarlon.drogaria.repository;
 import com.menezesmarlon.drogaria.domain.Categoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {
+public interface CategoriaRepository extends JpaRepository<Categoria, Short> {
 
 }
